client: hold reference for returned inode
CLient::make_request() returns pointer of target inode, but it does not
increase reference count of the returned inode. The inode may get freed
when CLient::make_request() releases MetaRequest.

The fix is hold a reference for the returned inode. Since there are many
of places use CLient::make_request() directly or indirectly, it's easy
to leak reference. This patch introduces a 'smart pointer' class and use
it to track the reference.

Signed-off-by: Yan, Zheng <zyan@redhat.com>